requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
Pillow==10.0.0
pycocotools==2.0.7
requests==2.31.0
torch==2.0.1
torchvision==0.15.2
